Build app with webpack + some perf improvements.
This change adds a package.json with webpack and associated dependencies to make it easier to build and run the demo app.

I also made a couple changes that should improve scrolling performance. There is an undocumented layer property, useBackingStore, which when used in moderation can significantly improve performance during scrolling. The reason it is not publicly exposed in React Canvas yet is that it's easy to abuse.

The backingStore property is useful for caching composite layouts (groups of text and/or image layers) so that they only need to be drawn once. The resulting bitmap is then later used when the layer needs to be redrawn. It's a little bit like setting translateZ(0) in CSS in that respect.

I've also improved image preloading by using the React Canvas ImageCache, rather than creating <img>'s directly. This should not be needed in future versions, but the code will still work.